Extract lens page shell helpers
