+++ CLAUDE.md
@@ -25,7 +25,7 @@ Web app (PWA) only: production runtime ships with no build system, no bundler, a
   - **Core**: `state.js` (single mutable `state`), `utils.js` (escape, status, notifications, dialogs), `theme.js`, `constants.js` (option arrays, personalities, EMF presets), `schema.js` (`MARKER_SCHEMA`, unit conversions, pricing, EMF thresholds, usage tracking), `main.js` (entry point, OAuth callback, init)
   - **Data pipeline**: `data.js` (`getActiveData` central pipeline), `profile.js` (CRUD + migrations), `adapters.js` (217 specialty markers across fattyAcids/metabolomix/oat/biostarks), `pdf-import.js`, `export.js`, `backup.js`, `crypto.js` (AES-256-GCM + BroadcastChannel cross-tab), `pii.js` (regex + local-AI obfuscation with streaming sanitizer)
   - **AI + providers**: `api.js` (6 providers, `callClaudeAPI` router, OpenRouter OAuth PKCE, Venice E2EE branch, recommended-model tiering), `provider-panels.js` (extracted from settings.js), `cashu-wallet.js` (Cashu BIP-39 seed, Lightning fund/withdraw, fee splitting), `nostr-discovery.js` (Routstr node discovery via Kind 38421)
-  - **Views + UI**: `views-router.js` (route validation, per-profile last-view restore, scroll anchoring), `lens-pages.js` (Labs/Genome/Body/Insight/Recommendations page renderers), `dashboard-widgets.js` (dashboard widget registry, defaults, per-profile widget prefs), `dashboard-widget-controls.js` (widget picker, layout actions, drag/reorder controls), `dashboard-widget-renderers.js` (dashboard widget body renderers and recommendation helpers), `category-view-renderers.js` (category chart cards, table/heatmap shells, fatty-acid profile renderers), `views.js` (dashboard composition, Light/category page orchestration, compare, correlations, detail modal, manual entry, focus card, onboarding), `nav.js` (sidebar, avatar colors), `charts.js` (Chart.js plugins), `context-cards.js` (9 cards with AI health dots + interpretive lens + change history), `client-list.js`, `supplements.js` (+ `supplement-warnings.js`, `food-contaminants.js`), `recommendations.js` (DNA-aware catalog, 3 touchpoints), `cycle.js`, `emf.js` (Baubiologie SBM-2015), `feedback.js`, `tour.js`, `changelog.js`, `notes.js`, `hardware.js` (GPU detection, model fitness), `image-utils.js`
+  - **Views + UI**: `views-router.js` (route validation, per-profile last-view restore, scroll anchoring), `lens-pages.js` (Labs/Genome/Body/Insight/Recommendations page renderers), `lens-page-shell.js` (shared lens header, widget chrome, ordering, and dashboard toggles), `dashboard-widgets.js` (dashboard widget registry, defaults, per-profile widget prefs), `dashboard-widget-controls.js` (widget picker, layout actions, drag/reorder controls), `dashboard-widget-renderers.js` (dashboard widget body renderers and recommendation helpers), `category-view-renderers.js` (category chart cards, table/heatmap shells, fatty-acid profile renderers), `views.js` (dashboard composition, Light/category page orchestration, compare, correlations, detail modal, manual entry, focus card, onboarding), `nav.js` (sidebar, avatar colors), `charts.js` (Chart.js plugins), `context-cards.js` (9 cards with AI health dots + interpretive lens + change history), `client-list.js`, `supplements.js` (+ `supplement-warnings.js`, `food-contaminants.js`), `recommendations.js` (DNA-aware catalog, 3 touchpoints), `cycle.js`, `emf.js` (Baubiologie SBM-2015), `feedback.js`, `tour.js`, `changelog.js`, `notes.js`, `hardware.js` (GPU detection, model fitness), `image-utils.js`
   - **Chat + context**: `chat.js` (panel, streaming, personalities, threads, images), `lab-context.js` (extracted: context assembly + memoization), `markdown.js` (extracted: XSS-safe rendering)
   - **Sync + agent access**: `sync.js` (Evolu CRDT — Phase 1 dual-write across 44 importedData surfaces via 3 planners + Phase 2 cutover gate; debounced `onDataSaved`; header badge). User-facing UI labels in the Diagnose modal: Phase 1 → "Push efficiency", Phase 2 → "Lean sync mode". `settings.js` (modal + sync setup), `data-merge.js` (composite-keyed conflict resolution + 200-cap on changeHistory, exports `COMPOSITE_KEYED_ARRAYS`)
   - **Wearables**: 7 vendors — Oura · Ultrahuman · WHOOP · Fitbit · Withings · Polar · Apple Health + `manual` as a first-class pseudo-source (user-authored weight/BP/pulse, migrated from legacy `importedData.biometrics`). `wearable-adapters.js` (canonical metric registry + per-vendor endpoint map + primary-source resolver; Apple Health + `manual` are `authType !== 'oauth2'`). Per-vendor pairs: `wearables-oura-auth.js` + `wearables-oura.js`, `wearables-withings-auth.js` + `wearables-withings.js`, `wearables-ultrahuman-auth.js` + `wearables-ultrahuman.js`, `wearables-whoop-auth.js` + `wearables-whoop.js`, `wearables-fitbit-auth.js` + `wearables-fitbit.js`, `wearables-polar-auth.js` + `wearables-polar.js`, `wearables-apple-health.js` (file-import, no OAuth), `wearables-manual.js` (no OAuth — `logManualMetric` / `logManualBP` / `deleteManualMetric` / `MANUAL_TAGS` whitelist / `migrateBiometricsToManual` idempotent migration). Server-side OAuth2 (secrets in `.env.local` + proxy): Oura / Withings / Ultrahuman / Polar. PKCE (no secret): WHOOP / Fitbit. Self-hosters override the hardcoded `clientId` per-vendor via `*_CLIENT_ID` env vars (browser fetches via `/api/proxy {wearable_runtime_config: true}` at startup; memoized + 1.5s timeout; scheduler awaits before first sync; empty env = no-op for hosted). Polar-only quirks: one-time `POST /v3/users` registration + transactions-model data reads (open → read → commit-only-after-IDB-write) — orchestrated by `postConnect` + `commitAfterWrite` hooks in `OAUTH_DISPATCH`. `wearables-store.js` (per-profile IndexedDB `labcharts-wearables-{profileId}` + `deleteDaily` for single-row cleanup + two-phase `upsertDailyBatch` (read-merge-put) so partial-fetch syncs don't null existing fields, never syncs), `wearables-summary.js` (L2 derivation + write gate — 5% d7 shift, trend flips, ISO-week-keyed rollovers, 14d force-refresh, plus `metric-removed` + `source-flip` + `latest-advanced` triggers so deletions and fresh data points propagate; `force` flag bypasses the gate for user-driven manual syncs), `wearables-connect.js` (connect/disconnect/backfill/sync orchestration + unified `OAUTH_DISPATCH` table), `wearables.js` (dashboard strip render + per-metric source picker + per-card "as of {date}" staleness hint + dismissible "Connect a wearable" stub for users with labs but no wearables + empty-state cards with inline log forms + detail-modal Manual entries list with per-row delete/backfill + ⇄ reorder mode persisting `importedData.wearableCardOrder` + Settings → Wearables list — theme-aware logos via brand-assets.js). `brand-assets.js` (per-vendor brand-asset registry — `iconLight/Dark` for in-app row icons, `signInLight/Dark` for landing-site Connect buttons, `mono` fallback). Brand asset files live in `brands/<vendor>/` with per-vendor LICENSE.md; mirrored to landing site via `brands/sync-to-site.sh`. WHOOP + Ultrahuman currently gated to "waiting on partner credentials" via `clientId: 'REPLACE_WITH_*'`. Polar logo also gated to fallback render until written-consent ticket lands. Sync deliberately strips `wearableConnections` from the synced payload (`sync.js:buildSyncPayload`) — refresh tokens stay local; users connect each vendor per-device.
