+++ tests/test-correctness-phase2.js
@@ -78,6 +78,7 @@ const pwaAppShellAssets = [
   '/js/data-merge.js',
   '/js/views-router.js',
   '/js/lens-pages.js',
+  '/js/lens-page-shell.js',
   '/js/dashboard-widgets.js',
   '/js/dashboard-widget-controls.js',
   '/js/dashboard-widget-renderers.js',
